Support Allies tray 4
Note, I didn't actually program what I needed for this tray as it was so custom. I combined the parts in TinkerCad.

It certainly could be programmed to look the same but I just didn't feel like taking the time for it. If it comes up in other expansions and I support mixing cards + other stuff, I'll update Allies tray 4 to match.